+++ javav2/example_code/glue/src/test/java/GlueTest.java
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import com.example.glue.*;
+import com.example.glue.scenario.GlueScenario;
 import com.google.gson.Gson;
 import software.amazon.awssdk.auth.credentials.EnvironmentVariableCredentialsProvider;
 import software.amazon.awssdk.services.glue.GlueClient;
@@ -10,9 +10,12 @@ import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient;
 import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueRequest;
 import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueResponse;
+
 import java.io.*;
 import java.net.URISyntaxException;
 import java.util.concurrent.TimeUnit;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 /**
  * To run these integration tests, you must set the required values
@@ -41,11 +44,11 @@ public class GlueTest {
     private static String bucketNameSc = "";
 
     @BeforeAll
-    public static void setUp() throws IOException, URISyntaxException {
+    public static void setUp() {
         glueClient = GlueClient.builder()
-                .region(Region.US_EAST_1)
-                .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
-                .build();
+            .region(Region.US_EAST_1)
+            .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
+            .build();
 
         // Get the values to run these tests from AWS Secrets Manager.
         Gson gson = new Gson();
@@ -61,86 +64,148 @@ public class GlueTest {
         existingDatabaseName = values.getExistingDatabaseName();
         existingCrawlerName = values.getExistingCrawlerName();
         jobNameSc = values.getJobNameSc() + java.util.UUID.randomUUID();
-        ;
         s3PathSc = values.getS3PathSc() + java.util.UUID.randomUUID();
-        ;
         dbNameSc = values.getDbNameSc() + java.util.UUID.randomUUID();
         crawlerNameSc = values.getCrawlerNameSc() + java.util.UUID.randomUUID();
         scriptLocationSc = values.getScriptLocationSc();
         locationUri = values.getLocationUri();
         bucketNameSc = values.getBucketNameSc();
-
-        // Uncomment this code block if you prefer using a config.properties file to
-        // retrieve AWS values required for these tests.
-        /*
-         * 
-         * try (InputStream input =
-         * GlueTest.class.getClassLoader().getResourceAsStream("config.properties")) {
-         * Properties prop = new Properties();
-         * if (input == null) {
-         * System.out.println("Sorry, unable to find config.properties");
-         * return;
-         * }
-         * prop.load(input);
-         * crawlerName = prop.getProperty("crawlerName");
-         * s3Path = prop.getProperty("s3Path");
-         * cron = prop.getProperty("cron");
-         * IAM = prop.getProperty("IAM");
-         * databaseName = prop.getProperty("databaseName");
-         * tableName = prop.getProperty("tableName");
-         * text = prop.getProperty("text");
-         * existingDatabaseName = prop.getProperty("existingDatabaseName");
-         * existingCrawlerName = prop.getProperty("existingCrawlerName");
-         * jobNameSc = prop.getProperty("jobNameSc")+ java.util.UUID.randomUUID();;
-         * s3PathSc = prop.getProperty("s3PathSc")+ java.util.UUID.randomUUID();;
-         * dbNameSc = prop.getProperty("dbNameSc")+ java.util.UUID.randomUUID();
-         * crawlerNameSc = prop.getProperty("crawlerNameSc")+
-         * java.util.UUID.randomUUID();
-         * scriptLocationSc = prop.getProperty("scriptLocationSc");
-         * locationUri = prop.getProperty("locationUri");
-         * bucketNameSc = prop.getProperty("bucketNameSc");
-         * 
-         * } catch (IOException ex) {
-         * ex.printStackTrace();
-         * }
-         */
     }
 
     @Test
     @Tag("IntegrationTest")
     @Order(1)
-    public void ScenarioTest() throws InterruptedException {
-        GlueScenario.createDatabase(glueClient, dbNameSc, locationUri);
-        GlueScenario.createGlueCrawler(glueClient, IAM, s3PathSc, cron, dbNameSc, crawlerNameSc);
-        GlueScenario.getSpecificCrawler(glueClient, crawlerNameSc);
-        GlueScenario.startSpecificCrawler(glueClient, crawlerNameSc);
-        GlueScenario.getSpecificDatabase(glueClient, dbNameSc);
+    void testCreateDatabase() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.createDatabase(glueClient, dbNameSc, locationUri);
+        });
+    }
 
-        System.out.println("Wait 5 min for the tables to become available");
-        TimeUnit.MINUTES.sleep(5);// Sleep for 5 minute to get tables ready
-        String myTableName = GlueScenario.getGlueTables(glueClient, dbNameSc);
-        GlueScenario.createJob(glueClient, jobNameSc, IAM, scriptLocationSc);
-        GlueScenario.startJob(glueClient, jobNameSc, dbNameSc, myTableName, bucketNameSc);
-        GlueScenario.getAllJobs(glueClient);
-        GlueScenario.getJobRuns(glueClient, jobNameSc);
-        GlueScenario.deleteJob(glueClient, jobNameSc);
-        System.out.println("*** Wait 5 MIN for the " + crawlerNameSc + " to stop");
-        TimeUnit.MINUTES.sleep(5);
-        GlueScenario.deleteDatabase(glueClient, dbNameSc);
-        GlueScenario.deleteSpecificCrawler(glueClient, crawlerNameSc);
+    @Test
+    @Tag("IntegrationTest")
+    @Order(2)
+    void testCreateGlueCrawler() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.createGlueCrawler(glueClient, IAM, s3PathSc, cron, dbNameSc, crawlerNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(3)
+    void testGetSpecificCrawler() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.getSpecificCrawler(glueClient, crawlerNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(4)
+    void testStartSpecificCrawler() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.startSpecificCrawler(glueClient, crawlerNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(5)
+    void testGetSpecificDatabase() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.getSpecificDatabase(glueClient, dbNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(6)
+    void testGetTable() {
+        assertDoesNotThrow(() -> {
+            System.out.println("*** Wait 5 min for the tables to become available");
+            TimeUnit.MINUTES.sleep(5);
+            System.out.println("6. Get tables.");
+            GlueScenario.getGlueTables(glueClient, dbNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(7)
+    void testCreateJob() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.createJob(glueClient, jobNameSc, IAM, scriptLocationSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(8)
+    void testStartJob() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.startJob(glueClient, jobNameSc, dbNameSc, tableName, bucketNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(9)
+    void testGetJobs() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.getAllJobs(glueClient);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(10)
+    void testRunJobs() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.getJobRuns(glueClient, jobNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(11)
+    void testDeleteJob() {
+        assertDoesNotThrow(() -> {
+            GlueScenario.deleteJob(glueClient, jobNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(12)
+    void testDeleteDB() {
+        assertDoesNotThrow(() -> {
+            System.out.println("*** Wait 5 MIN for the " + crawlerNameSc + " to stop");
+            TimeUnit.MINUTES.sleep(5);
+            GlueScenario.deleteDatabase(glueClient, dbNameSc);
+        });
+    }
+
+    @Test
+    @Tag("IntegrationTest")
+    @Order(13)
+    void testDelCrawler() {
+        assertDoesNotThrow(() -> {
+            System.out.println("*** Wait 5 MIN for the " + crawlerNameSc + " to stop");
+            TimeUnit.MINUTES.sleep(5);
+            GlueScenario.deleteSpecificCrawler(glueClient, crawlerNameSc);
+        });
     }
 
     private static String getSecretValues() {
-        // Get the Amazon RDS creds from Secrets Manager.
         SecretsManagerClient secretClient = SecretsManagerClient.builder()
-                .region(Region.US_EAST_1)
-                .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
-                .build();
+            .region(Region.US_EAST_1)
+            .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
+            .build();
         String secretName = "test/glue";
 
         GetSecretValueRequest valueRequest = GetSecretValueRequest.builder()
-                .secretId(secretName)
-                .build();
+            .secretId(secretName)
+            .build();
 
         GetSecretValueResponse valueResponse = secretClient.getSecretValue(valueRequest);
         return valueResponse.secretString();
